+++ booleancondition.h
@@ -40,7 +40,7 @@ public:
     void setValue(qint64);
     QString toString();
 
-    virtual quint8 getValidRangeSize(quint64 faces) const;
+    virtual quint64 getValidRangeSize(quint64 faces) const;
 
 private:
     LogicOperator m_operator;
